requirements.txt
Django==4.1.2
asgiref
dj-database-url
django-cors-headers
django-discover-runner
django-environ
django-role-permissions
djangorestframework
djangorestframework-simplejwt
gunicorn
mysqlclient==2.1.1
PyJWT
pytz
six
sqlparse
uuid
whitenoise
django-storages
boto3
django-cors-headers
untangle
requests
beautifulsoup4
django-ses
xmltodict
PyPDF2
twilio
filetype
drf-extra-fields
djangorestframework-xml
pandas
xarray
openpyxl
pdfkit
django-crontab